Fixed #9479([German UI] Improved main menu font) Patch by einheit-101
